+++ docs/src/tutorials/calibration/obs_site_level_calibration.jl
@@ -180,8 +180,16 @@ function model(Vcmax25, g1)
     )
 
     #md # Create integrated land model
-    land_model =
-        LandModel{FT}(forcing, LAI, toml_dict, domain, Δt; soil, canopy)
+    land_model = LandModel{FT}(
+        forcing,
+        LAI,
+        toml_dict,
+        domain,
+        Δt;
+        prognostic_land_components,
+        soil,
+        canopy,
+    )
 
     #md # Set initial conditions from FLUXNET data
     set_ic! = FluxnetSimulations.make_set_fluxnet_initial_conditions(